fcft: don't get stuck on invalid fontfeatures
If hb_feature_from_string() failed, we ended up in an infinite loop.

First, we didn't break out of the loop when
hb_feature_from_string(). But, we also did not increase the
font->hb_feats counter. This meant the next iteration of the loop
tried the exact same fontfeature pattern again, and again and again...

Fix by using a separate counter for the fontconfig pattern index. This
way, we can iterate all fontfeatures patterns, while skipping the
invalid ones.

Closes foot-1554